revert creation of a __main__.py
