+++ setup.cfg
@@ -25,4 +25,4 @@ include_package_data=True
 
 [options.entry_points]
 console_scripts =
-    hazen = hazenlib:__main__.main
+    hazen = hazenlib:main
